[Fix] Option의 info data를 관리하는 파일을 생성
